New Binning for Flavor Effs. and Datasets ran with New Medium Working Point
Also, datasets plotted in Preliminary Analysis notebook to check the size/shape of datasets.  It appears that ttbar and QCD should be normalized using sumw2 with the inclusion of their respective cross sections, while the RSgluon and DM samples should use the total number of generated events without the cross sections.