Saving Flavor Efficiency Uproot Job Mow Creates the LUTs Necessary for btag systematics with --useEff Option
Also, the BTagUpdater had to be slightly updated, as the only tests from several months ago simply checked that the function didn't break the processor.  Now, it's written to work.  The main changes were to flatten the boolean array that defines conditions 1-4, in order to match the BSF array, as well as return the properly shaped boolean for defining btags.  The other main change was to specifically call the 'efficiency' values from the pandas dataframes, which was a step that was missing for some reason.  Finally, the subjet pt and eta must be restricted to the bin sizes defined in the correction files.