Included More Cutflow Information to Examine How Many Jets Are Included in First Bin
Candan and I [AC] suspected that the number of events in this bin are too insignificant to be included, hence the bin size here should be larger, but making it any larger will lose the beginning of the "turn on" curve, making that turn-on JetHT value more ambiguous.  However, this turn-on value is not as important as the value which the efficiency begins to saturate, so this may not truly be an issue, but a nitpick.