Modified Code to be Able to Run Flavor Efficiency Analysis and Save Efficiency Look-Up Tables in New Directory Naming Convention
Removed the, now redundant, old naming convention of the efficiency LUTs.  This is because it has already been decided that the best courses of action will be to accept that the efficiency bins must be rather large compared with the 2016 AN results.  Therefore, I do not specify the size of the bins in the name of the .csv file anymore.