Update README.md
A test for including entire pdf's in the readme instructions.  For this test, I have included a link to download a pdf that gives a brief introduction to start using Coffea-Casa with an analysis.  I can make an altered version of this pdf to only give instructions explicitly for using this repository, but for now, it is self-explanatory (and better than nothing).  Ultimately, I want the images of the pdf slides to be displayed right here in the readme file, rather than be a separate download link.