First Test ROOT Files Made with BTag Systematics and Top pT Re-weighting
The naming convention may change in the future, as this is the first naming convention that came to mind.  However, this format is most likely not the recommended.  Therefore, this commit serves as the first of many tests to come in producing root files with the necessary analysis corrections to the background estimate.