First Batch of Weighted Coffea Output Files with bTagSyst Applied Using MC Flavor Efficiency Look Up Tables
This method updates the status of the btag boolean for each subjet per event.  The fraction of events to be updated (promoted or demoted) is determined by the btagging efficiency of the subjets as a function of the corresponding subjet's pt and eta.  This is the method that was used for the 2016 AN and may continue to be the method we move along with.  However, the results from these systematics/uncertainties should  be compared with "method2" output files to determine which is better suited for the analysis.