Some Histogram Ranges Will Now Exclude 'Cut' Regions, and Other Subtle Changes
In the second uproot job the subtle changes are referring to how some unnecessary event weights were being applied.  The mistag rate event weight was being applied to all MC signals beyond just the data and ttbar MC for the sake of generating a background estimate.  Also, the mass modification procedure is now only being done to the data for the sake of correcting the jet mass in the pretag category for making a better background estimate.