Updated Coffea Outputs with Smarter Binning and Removed Signal MC Weighted Redundancy
Also, these outputs were made with the sumw and sumw2 being defined in the proper place in the workflow, before any cuts to the data could be made, to ensure that the values were accurate for the purpose of normalization.

Now, the coffea outputs for the systematics must be made, as well as those for the other provided working point, as these are all done with the loose working point.