Potentially Dumb Question, But I'm Curious...
Even though we are selecting every third root file, that does not mean that we are taking exactly 1/3 of the available data, unless every root file has the same number of events.  Upon checking the cutflow, it looks like we took about 1/15 of the data.  This would mean that we only scale the ttbar MC to 1/15th of the total 2016 Luminosity (instead of a third of the luminosity), correct?  This allows for much better agreement between bkg. est. and data, but only predicts several Tops given the small luminosity.