New Trigger Analysis Coffea Outputs
Also, there was a logical error in our older version of the UL20 2016 trigger analysis.  The plots that should've showed the efficiencies for all passable events, despite the mass window, used to only plot events that explicitly did not pass the mass window requirement.  This has been fixed for 2016 and is implemented in 2017.  2018 must be run and saved.