Background Estimation with Individual Eras and the Total of the Individual Eras
This is done in light of seeing questionable results from processing several eras at a time into a single output.  The attempted solution here was to process the eras as their own individual outputs with individual mistag rates, and carefully scale and combine the histograms to get a better NTMJ estimate.  I believe this way yields better results from the former method of processing them all at once in a single output script.