Tested small recipe for saving new mistag rates in notebook after checking them
Will work on making this into a new script that will replace the old one as the default way to make lookup tables for the mistag rates.  Will also work on making a script that adds the eras of a given year together to avoid copying and pasting large chunks of code over and over for this process and several others.